add:增加robot_service对电机矢量的缩放和偏执

+++ user/robot/platform/inc/platform.h
@@ -3,7 +3,7 @@
 #include<stdio.h>
 
 //axis 的精度
-typedef float AxisFloat;
+typedef double AxisFloat;
 //
 
 extern void* (*robotMalloc)(size_t size);
